pyproject.toml
[tool.poetry]
name = "simple-bot"
version = "0.1.0"
description = ""
authors = []
license = ""

[tool.poetry.dependencies]
python = "^3.9"
vkbottle = "^3.0.2"
envparse = "^0.2.0"
tortoise-orm = "^0.17.8"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
